[eInternetCheck] * provide async internet check
