+++ lib/base/Makefile.inc
@@ -17,6 +17,7 @@ base_libenigma_base_a_SOURCES = \
 	base/freesatv2.cpp \
 	base/huffman.cpp \
 	base/init.cpp \
+	base/internetcheck.cpp \
 	base/ioprio.cpp \
 	base/message.cpp \
 	base/modelinformation.cpp \
@@ -53,6 +54,7 @@ baseinclude_HEADERS = \
 	base/itssource.h \
 	base/init.h \
 	base/init_num.h \
+	base/internetcheck.h \
 	base/ioprio.h \
 	base/message.h \
 	base/modelinformation.h \
